requirements.txt
asgiref==3.4.1
black==21.11b1
click==8.0.3
Django==3.2.9
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.1
